[CM][Code Clean-up] Fix issues reported by cppchecker.

+++ media_driver/agnostic/common/cm/cm_hal_dump.cpp
@@ -156,8 +156,10 @@ int32_t HalCm_DumpCommadBuffer(PCM_HAL_STATE state, PMOS_COMMAND_BUFFER cmdBuffe
 
     // Set the file name. linux
     //PlatformSNPrintf(fileName, MOS_MAX_HLT_FILENAME_LEN, HALCM_COMMAND_BUFFER_OUTPUT_DIR);
-    PlatformSNPrintf(fileName + strlen(fileName), MOS_MAX_HLT_FILENAME_LEN - strlen(fileName), PLATFORM_DIR_SEPERATOR);
-    PlatformSNPrintf(fileName + strlen(fileName), MOS_MAX_HLT_FILENAME_LEN - strlen(fileName), "%s_%d.txt", HALCM_COMMAND_BUFFER_OUTPUT_FILE, commandBufferNumber);
+    PlatformSNPrintf(fileName + strlen(fileName), MOS_MAX_HLT_FILENAME_LEN - strlen(fileName),
+                     PLATFORM_DIR_SEPERATOR);
+    PlatformSNPrintf(fileName + strlen(fileName), MOS_MAX_HLT_FILENAME_LEN - strlen(fileName),
+                     "%s_%d.txt", HALCM_COMMAND_BUFFER_OUTPUT_FILE, (int)commandBufferNumber);
 
     numberOfDwords = cmdBuffer->iOffset / sizeof(uint32_t);
 
@@ -287,8 +289,10 @@ int32_t HalCm_DumpCurbeData(PCM_HAL_STATE state)
     GetLogFileLocation(HALCM_CURBE_DATA_OUTPUT_DIR, fileName);
 
    // PlatformSNPrintf(fileName, MOS_MAX_HLT_FILENAME_LEN, HALCM_CURBE_DATA_OUTPUT_DIR);
-    PlatformSNPrintf(fileName + strlen(fileName), MOS_MAX_HLT_FILENAME_LEN - strlen(fileName), PLATFORM_DIR_SEPERATOR);
-    PlatformSNPrintf(fileName + strlen(fileName), MOS_MAX_HLT_FILENAME_LEN - strlen(fileName), "%s_%d.txt", HALCM_CURBE_DATA_OUTPUT_FILE, curbeDataNumber);
+    PlatformSNPrintf(fileName + strlen(fileName), MOS_MAX_HLT_FILENAME_LEN - strlen(fileName),
+                     PLATFORM_DIR_SEPERATOR);
+    PlatformSNPrintf(fileName + strlen(fileName), MOS_MAX_HLT_FILENAME_LEN - strlen(fileName),
+                     "%s_%d.txt", HALCM_CURBE_DATA_OUTPUT_FILE, (int)curbeDataNumber);
 
     // write curbe data dwords.
     if (state->dshEnabled)
